feat: 动态价格计算

+++ packages/client/src/pages/ShoppingCart/index.tsx
@@ -2,11 +2,11 @@ import React, { FC, useMemo, useState } from 'react';
 import AdminContainer from '@src/components/biz/AdminContainer';
 import styles from './style.less';
 import { columns } from '@src/pages/ShoppingCart/helper';
-import { Affix, Button, message, Table } from 'antd';
+import { Affix, Button, message, Modal, Table } from 'antd';
 import AdminTitleBar from '@src/components/biz/AdminTitleBar';
 import { usePersistFn } from 'ahooks';
 import useMountRequest from '@src/pages/ShoppingCart/useHooks/useMountRequest';
-import { map, sum } from 'lodash';
+import { map, sum, filter, includes } from 'lodash';
 import produce from 'immer';
 import { deleteShoppingCartRequest, updateShoppingCartRequest } from '@src/service';
 
@@ -44,18 +44,34 @@ const ShoppingCart: FC = () => {
   });
 
   const handleTotalPrice = useMemo(() => {
-    const valueList = map(selectedRowState, item => {
-      return item.count * item.price;
+    const selectedRow = filter(list, item => includes(selectedRowKeys, item.id));
+    const valueList = map(selectedRow, item => {
+      return item?.count * item?.price;
     });
 
     return sum(valueList);
-  }, [selectedRowState?.length]) as number;
+  }, [list, selectedRowKeys]) as number;
 
-  const handleDelete = usePersistFn(async () => {
-    console.log(selectedRowKeys);
-    await deleteShoppingCartRequest(selectedRowKeys);
-    await refresh();
-    message.success('删除成功');
+  const handleDeleteAll = usePersistFn(async () => {
+    Modal.confirm({
+      title: '确认删除选中的所有玩具商品?',
+      onOk: async () => {
+        await deleteShoppingCartRequest(selectedRowKeys);
+        await refresh();
+        message.success('删除成功');
+      },
+    });
+  });
+
+  const handleDeleteItem = usePersistFn(async (id: number) => {
+    Modal.confirm({
+      title: '确认删除当前玩具商品？',
+      onOk: async () => {
+        await deleteShoppingCartRequest([id]);
+        await refresh();
+        message.success('删除成功');
+      },
+    });
   });
 
   return (
@@ -68,7 +84,7 @@ const ShoppingCart: FC = () => {
             pagination={false}
             loading={loading}
             rowSelection={rowSelection}
-            columns={columns({ handleChangeCount })}
+            columns={columns({ handleChangeCount, handleDeleteItem })}
             dataSource={map(list, item => Object.assign({}, item, { key: item.id }))}
           />
         </div>
@@ -76,7 +92,7 @@ const ShoppingCart: FC = () => {
 
       <Affix offsetBottom={12}>
         <footer className={styles.settlement}>
-          <div>{selectedRowKeys.length > 0 && <a onClick={handleDelete}>删除全部</a>}</div>
+          <div>{selectedRowKeys.length > 0 && <a onClick={handleDeleteAll}>删除全部</a>}</div>
           <div className="handle-settlement">
             <p>
               已选择商品 <span className="count">{selectedRowKeys.length}</span> 件
